pyproject.toml
[project]
name = "02-openrouter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.18",
    "python-dotenv>=1.1.0",
]
